add task2

+++ HW1/task1.hs
@@ -1,4 +1,4 @@
-f :: Double -> Double
+f :: Integer -> Double
 f n = f' n 1
 f' 1 p = 1 + 1 / p
 f' n p = f' (n - 1) (1 + 1 / p)
